# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Youth CapitalCore API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /press:
    get:
      operationId: getArticles
      tags: [press]
      summary: List news articles
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [simulation, platform]
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Articles list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleListResponse"
    post:
      operationId: createArticle
      tags: [press]
      summary: Create a news article (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArticleRequest"
      responses:
        "201":
          description: Article created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
  /press/{id}:
    get:
      operationId: getArticle
      tags: [press]
      summary: Get a single article by slug (or legacy numeric id)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Article details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        fullName:
          type: string
        fullNameAr:
          type: string
        role:
          type: string
          enum: [user, admin]
        simulationRole:
          type: string
          nullable: true
        region:
          type: string
          nullable: true
        languagePreference:
          type: string
          enum: [en, ar]
        bio:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, banned]
        applicationStatus:
          type: string
          enum: [none, pending, approved, rejected]
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, email, fullName, role, languagePreference, status]
...
    Article:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        title:
          type: string
        titleAr:
          type: string
        content:
          type: string
        contentAr:
          type: string
        type:
          type: string
          enum: [simulation, platform]
        author:
          $ref: "#/components/schemas/User"
        thumbnailUrl:
          type: string
          nullable: true
        publishedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, slug, title, titleAr, content, contentAr, type, author, publishedAt, createdAt]
    ArticleListResponse:
      type: object
      properties:
        articles:
          type: array
          items:
            $ref: "#/components/schemas/Article"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [articles, total]
    CreateArticleRequest:
      type: object
      properties:
        slug:
          type: string
          description: Optional custom URL slug. Auto-generated from the title when omitted.
        title:
          type: string
        titleAr:
          type: string
        content:
          type: string
        contentAr:
          type: string
        type:
          type: string
          enum: [simulation, platform]
        thumbnailUrl:
          type: string
          nullable: true
      required: [title, titleAr, content, contentAr, type]
...