# packages/api-zod/src/generated/types/articleListResponse.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Api
 * Youth CapitalCore API
 * OpenAPI spec version: 0.1.0
 */
import type { Article } from "./article";

export interface ArticleListResponse {
  articles: Article[];
  total: number;
  page?: number;
  limit?: number;
}