# packages/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Api
 * Youth CapitalCore API
 * OpenAPI spec version: 0.1.0
 */
import type { UserApplicationStatus } from "./userApplicationStatus";
import type { UserLanguagePreference } from "./userLanguagePreference";
import type { UserRole } from "./userRole";
import type { UserStatus } from "./userStatus";

export interface User {
  id: number;
  email: string;
  fullName: string;
  fullNameAr?: string;
  role: UserRole;
  simulationRole?: string | null;
  region?: string | null;
  languagePreference: UserLanguagePreference;
  bio?: string | null;
  status: UserStatus;
  applicationStatus?: UserApplicationStatus;
  createdAt?: Date;
}