# packages/db/src/schema/press.ts
import { pgTable, text, serial, timestamp, pgEnum, integer } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod";
import { usersTable } from "./users.js";

export const articleTypeEnum = pgEnum("article_type", ["simulation", "platform"]);

export const articlesTable = pgTable("articles", {
  id: serial("id").primaryKey(),
  // Nullable for now: existing rows need a one-time backfill (see
  // scripts/src/backfill-article-slugs.ts) before this can become NOT NULL.
  // New rows always get one from the create-article route.
  slug: text("slug").unique(),
  title: text("title").notNull(),
  titleAr: text("title_ar").notNull(),
  content: text("content").notNull(),
  contentAr: text("content_ar").notNull(),
  type: articleTypeEnum("type").notNull(),
  authorId: integer("author_id").notNull().references(() => usersTable.id),
  thumbnailUrl: text("thumbnail_url"),
  publishedAt: timestamp("published_at").notNull().defaultNow(),
  createdAt: timestamp("created_at").notNull().defaultNow(),
});

export const insertArticleSchema = createInsertSchema(articlesTable).omit({ id: true, createdAt: true });
export type InsertArticle = z.infer<typeof insertArticleSchema>;
export type Article = typeof articlesTable.$inferSelect;